+++ spec/spec_helper.rb
@@ -7,7 +7,10 @@ require_relative "../lib/lumberjack_sidekiq"
 
 RSpec.configure do |config|
   config.warnings = true
+  config.disable_monkey_patching!
+  config.default_formatter = "doc" if config.files_to_run.one?
   config.order = :random
+  Kernel.srand config.seed
 end
 
 class MyWorker